# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ms-customers documentation",
        "description": "Customers ms, api laravel 10",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "api/customers/list": {
            "get": {
                "tags": [
                    "customers"
                ],
                "summary": "List Customers",
                "operationId": "763e4bbf3b1c5962c540cad4968bbc21",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Rodolfo molina"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "3012074828"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customers/register": {
            "post": {
                "tags": [
                    "customers"
                ],
                "summary": "Register customers",
                "operationId": "f7e3365a3c515f099f4e1a9555863850",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "canela molina",
                                    "phone": "3012094939"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "canela molina"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "3012094939"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "put": {
                "tags": [
                    "customers"
                ],
                "summary": "Update customers",
                "operationId": "32da1527afa82d1a6b0e034146f1032d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "canela molina Editado",
                                    "phone": "77777777"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "canela molina Editado"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "777777777"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "customers"
                ],
                "summary": "Delete customers",
                "operationId": "e9f16ac94d63309225711a1f3dfb245a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo realizar correctamente la operación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customers/search/{id}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "summary": "Search customers list",
                "operationId": "aedf25fe00732c6747b2c748302e2448",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Rodolfo molina"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "3012074828"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Customer] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}